+++ tests/unit/executor/test_basic.py
@@ -1,5 +1,7 @@
 import pytest
 import inspect
+from tartiflette.resolver.factory import default_error_coercer
+from tartiflette.types.exceptions.tartiflette import GraphQLError
 from unittest.mock import Mock
 
 
@@ -85,8 +87,29 @@ async def test_executor_basic_execute(errors, expected, monkeypatch):
 
     from tartiflette.executors.basic import execute
 
-    a = await execute([], request_ctx={})
+    a = await execute([], request_ctx={}, error_coercer=default_error_coercer)
 
     assert a == expected
+
+    monkeypatch.undo()
+
+
+@pytest.mark.asyncio
+async def test_executor_basic_execute_custom_resolver(monkeypatch):
+    from tartiflette.executors import basic
+
+    async def my_execute(_, exec_ctx, __):
+        exec_ctx.add_error(GraphQLError("My error"))
+
+    def custom_error_coercer(exception):
+        return {"message": "Error from custom_error_coercer"}
+
+    monkeypatch.setattr(basic, "_execute", my_execute)
+
+    from tartiflette.executors.basic import execute
+
+    a = await execute([], request_ctx={}, error_coercer=custom_error_coercer)
+
+    assert a == {"data": {}, "errors": [{"message": "Error from custom_error_coercer"}]}
 
     monkeypatch.undo()