+++ tests/unit/parser/test_parser.py
@@ -25,7 +25,7 @@ def test_tartiflette_request_parser(clean_registry, monkeypatch):
 
     trp = TartifletteRequestParser()
 
-    assert trp.parse_and_tartify(None, "query aq { __schema }") == []
+    assert trp.parse_and_tartify(None, "query aq { __schema }") == ([], None)
 
     monkeypatch.undo()
 
@@ -42,7 +42,6 @@ def test_tartiflette_request_parser_vistor_except(clean_registry, monkeypatch):
 
     trp = TartifletteRequestParser()
 
-    with pytest.raises(Exception):
-        trp.parse_and_tartify(None, "query aq { __schema }")
+    assert trp.parse_and_tartify(None, "query aq { __schema }") == ([], None)
 
     monkeypatch.undo()