+++ tests/unit/test_engine.py
@@ -1,4 +1,5 @@
 import pytest
+from tartiflette.resolver.factory import default_error_coercer
 from unittest.mock import Mock
 
 
@@ -35,23 +36,132 @@ async def test_engine_execute(clean_registry):
 
 
 @pytest.mark.asyncio
+async def test_engine_execute_parse_error(clean_registry):
+    from tartiflette.engine import Engine
+
+    e = Engine("type Query { a: String }")
+
+    assert await e.execute("query { unknownNode1 unknownNode2 }") == {
+        "data": None,
+        "errors": [
+            {
+                "message": "field `Query.unknownNode1` was not found in GraphQL schema.",
+                "path": ["unknownNode1"],
+                "locations": [
+                    {
+                        "column": 9,
+                        "line": 1,
+                    },
+                ],
+            },
+            {
+                "message": "field `Query.unknownNode2` was not found in GraphQL schema.",
+                "path": ["unknownNode1", "unknownNode2"],
+                "locations": [
+                    {
+                        "column": 22,
+                        "line": 1,
+                    },
+                ],
+            },
+        ],
+    }
+
+
+@pytest.mark.asyncio
+async def test_engine_execute_custom_error_coercer(clean_registry):
+    from tartiflette.engine import Engine
+
+    def custom_error_coercer(exception):
+        error = default_error_coercer(exception)
+        error["message"] = error["message"] + "Custom"
+        return error
+
+    e = Engine("type Query { a: String }", error_coercer=custom_error_coercer)
+
+    assert await e.execute("query { unknownNode1 unknownNode2 }") == {
+        "data": None,
+        "errors": [
+            {
+                "message": "field `Query.unknownNode1` was not found in GraphQL schema.Custom",
+                "path": ["unknownNode1"],
+                "locations": [
+                    {
+                        "column": 9,
+                        "line": 1,
+                    },
+                ],
+            },
+            {
+                "message": "field `Query.unknownNode2` was not found in GraphQL schema.Custom",
+                "path": ["unknownNode1", "unknownNode2"],
+                "locations": [
+                    {
+                        "column": 22,
+                        "line": 1,
+                    },
+                ],
+            },
+        ],
+    }
+
+
+@pytest.mark.asyncio
 async def test_engine_execute_empty_req_except(clean_registry):
     from tartiflette.engine import Engine
 
-    e = Engine("type Query { a:String }")
+    e = Engine("type Query { a: String }")
 
-    with pytest.raises(Exception):
-        await e.execute("")
+    assert await e.execute("") == {
+        "data": None,
+        "errors": [
+            {
+                "message": "1.1: syntax error, unexpected EOF",
+                "path": None,
+                "locations": [],
+            },
+        ],
+    }
 
 
 @pytest.mark.asyncio
 async def test_engine_execute_syntax_error(clean_registry):
     from tartiflette.engine import Engine
 
-    e = Engine("type Query { a:String }")
+    e = Engine("type Query { a: String }")
 
-    with pytest.raises(Exception):
-        await e.execute("query { a { }")
+    assert await e.execute("query { a { }") == {
+        "data": None,
+        "errors": [
+            {
+                "message": "1.12: unrecognized character \\xc2",
+                "path": None,
+                "locations": [],
+            },
+        ],
+    }
+
+
+@pytest.mark.asyncio
+async def test_engine_execute_unhandled_exception(clean_registry):
+    from tartiflette.engine import Engine
+
+    e = Engine("type Query { a: Ninja } type Ninja { a: String }")
+
+    assert await e.execute("""
+        fragment AFragment on Ninja { a }
+        fragment AFragment on Ninja { a }
+        query { a { } }
+    """) == {
+        "data": None,
+        "errors": [
+            {
+                "message": "4.20: unrecognized character \\xc2",
+                "path": None,
+                "locations": [],
+            },
+        ],
+    }
 
 
 @pytest.mark.asyncio